Added access to various MySQL constants, plus field flags. More specific exception handling. And a light sprinkling of bug fixes throughout.
I'm thinking this part is done, unless I find the need to add another
feature or two for the Python DBI. All the mysql_* calls are duplicated
either as functions, connection objects, or result objects, with the
exception of deprecated functions and a lot of field_* functions.
Access to the MYSQL_FIELD members is done mostly through
ResultObject.describe(), which returns almost all information for all
the fields. ResultObject.field_flags() returns the flag values.


git-svn-id: https://mysql-python.svn.sourceforge.net/svnroot/mysql-python@4 9fc6cd9a-920d-0410-adcf-ac96716ed7e8