Break up the various cursor variations into MixIn classes.
Found a work-around for the way MySQL treats TIME literals with
dates and/or fractional seconds.

Added a mutex into the cursor so that connections can be shared between
threads. threadsafety=2

I could easily make the cursors sharable as well (threadsafety=3) but
I hardly see the point. Even sharing connections is not a good idea,
because you don't get the benefit of multiple mysqld threads.


git-svn-id: https://mysql-python.svn.sourceforge.net/svnroot/mysql-python@42 9fc6cd9a-920d-0410-adcf-ac96716ed7e8